+++ onward/app/feed/MostPopularAgent.scala
@@ -187,77 +187,3 @@ class DayMostPopularAgent(contentApiClient: ContentApiClient, ophanApi: OphanApi
     }
   }
 }
-
-case class DeeplyReadItem(
-    path: String,
-    benchmarkedAttentionTime: Int,
-    url: String,
-    linkText: String,
-    showByline: Boolean,
-    byline: String,
-    image: String,
-    isLiveBlog: Boolean,
-    pillar: String,
-    designType: String,
-    webPublicationDate: String,
-    headline: String,
-    shortUrl: String,
-)
-object DeeplyReadItem {
-  implicit val jsonWrites = Json.writes[DeeplyReadItem]
-}
-
-class DeeplyReadAgent(contentApiClient: ContentApiClient, ophanApi: OphanApi) {
-
-  private val mapping: scala.collection.mutable.Map[String, Content] =
-    scala.collection.mutable.Map.empty[String, Content]
-
-  def refresh()(implicit ec: ExecutionContext): Future[Unit] = {
-    ophanApi.getDeeplyReadContent().map { seq =>
-      seq.foreach { i =>
-        val path = i.path
-        println(s"Looking up data for path: ${path}")
-        val capiItem = contentApiClient
-          .item(path)
-          .showTags("all")
-          .showFields("all")
-          .showReferences("all")
-          .showAtoms("all")
-        contentApiClient
-          .getResponse(capiItem)
-          .map { res =>
-            res.content.map { c =>
-              mapping += (path -> c)
-            }
-          }
-      }
-    }
-    Future.successful(())
-  }
-
-  def getDataForPath(path: String): Option[Content] = {
-    mapping.get(path)
-  }
-
-  def getDeeplyReadItemForOphanItem(item: OphanDeeplyReadItem): Option[DeeplyReadItem] = {
-    for {
-      content <- getDataForPath(item.path)
-      webPublicationDate <- content.webPublicationDate
-    } yield DeeplyReadItem(
-      path = item.path,
-      benchmarkedAttentionTime = item.benchmarkedAttentionTime,
-      url = content.webUrl,
-      linkText = "Brexit: No 10 refuses to commit to agreeing with EU's new no-deal contingency plan – live updates",
-      showByline = false,
-      byline = "Andrew Sparrow",
-      image =
-        "https://i.guim.co.uk/img/media/aa9685c6ddfa64e1844bee8d39c5d09864ef93c2/0_122_6048_3628/master/6048.jpg?width=300&quality=85&auto=format&fit=max&s=3f734551601ba14c590fdda3726c2334",
-      isLiveBlog = true,
-      pillar = "news",
-      designType = "Live",
-      webPublicationDate = webPublicationDate.toString(),
-      headline = "Brexit: No 10 refuses to commit to agreeing with EU's new no-deal contingency plan – live updates",
-      shortUrl = "https://gu.com/p/fyzj7",
-    )
-  }
-}
